+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Fix Russia calendar: non-working days are shifted to the next MON when they happen on the week-end (#589).
 - Fix Russia New year holidays. It has become a week off since 2005 (related to #578).
+- Added a `daterange` function in `workalendar.core` module to iterate between two dates.
 
 ## v13.0.0 (2020-11-13)