+++ Changelog.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master (unreleased)
+## v14.1.0 (2020-12-10)
 
 - Fix Russia 2021 holidays, thx @MichalChromcak for the bug report (#578).